+++ src/main/java/javax/util/streamex/StreamEx.java
@@ -791,6 +791,8 @@ public class StreamEx<T> extends AbstractStreamEx<T, StreamEx<T>> {
      */
     @SuppressWarnings("unchecked")
     public StreamEx<T> append(T... values) {
+        if(values.length == 0)
+            return this;
         return append(Stream.of(values));
     }
 
@@ -817,6 +819,8 @@ public class StreamEx<T> extends AbstractStreamEx<T, StreamEx<T>> {
      */
     @SuppressWarnings("unchecked")
     public StreamEx<T> prepend(T... values) {
+        if(values.length == 0)
+            return this;
         return prepend(Stream.of(values));
     }